fluor_yield.dat: K-shell fluorescence yields revision
Alex Simionovici noticed an unusual drop in the K-shell fluorescence
yields, which makes no sense physically.
This was already noted in Hubbell's 2004 erratum to his 1994 paper "A
review, bibliography, and tabulation of K, L, and higher atomic shell
X-ray fluorescence yields. Journal of physical and chemical reference
data, 23(2):339–364, 1994."

As recommended by him, the decision was made to use the
Bambynek/Cohen/Burhop/Hubbell fluoresence yields for the K-shell instead